Refactored Auth folder
